requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
openpyxl>=3.1.0
xlrd>=2.0.1
plotly>=5.18.0
sqlalchemy>=2.0.0
python-dateutil>=2.8.0
apscheduler>=3.10.0
